feat: add UserAuthenError
